@@ -16,15 +16,16 @@
   "scripts": {
     "build": "rm -rf lib && tsc --outDir lib && cp flowtypes/* lib && rm lib/*.test.*",
     "build-storybook": "build-storybook -s public && rm ./examples/*.d.ts && rm ./src/*.d.ts",
-    "jest:e2e": "jest --runInBand e2e",
+    "jest:e2e": "jest --runInBand",
     "jest:e2e:update": "jest --runInBand -u e2e",
     "storybook": "start-storybook -p 9010 -s public",
     "storybook:ci": "start-storybook -p 9011 -s public --ci --quiet",
-    "test": "yarn test:e2e && yarn test:unit",
+    "test": "yarn typecheck && yarn test:e2e && yarn test:unit",
     "test:e2e": "start-server-and-test storybook:ci http-get://localhost:9011 jest:e2e",
     "test:e2e:dev": "E2E=dev yarn jest:e2e",
     "test:e2e:update": "start-server-and-test storybook:ci http-get://localhost:9011 jest:e2e:update",
-    "test:unit": "jest src"
+    "test:unit": "jest src",
+    "typecheck": "tsc --noEmit"
   },
   "tags": [
     "react"
@@ -45,28 +46,27 @@
     "movable"
   ],
   "devDependencies": {
-    "@storybook/addon-options": "4.2.0-alpha.8",
-    "@storybook/addons": "4.2.0-alpha.8",
-    "@storybook/react": "4.2.0-alpha.8",
-    "@types/expect-puppeteer": "^3.3.0",
-    "@types/jest": "^23.3.13",
-    "@types/jest-environment-puppeteer": "^2.2.1",
-    "@types/jest-image-snapshot": "^2.4.0",
-    "@types/node": "^10.12.18",
-    "@types/puppeteer": "^1.11.2",
-    "@types/react": "^16.7.20",
-    "@types/react-dom": "^16.0.11",
-    "@types/storybook__react": "^4.0.0",
+    "@storybook/react": "5.2.8",
+    "@types/expect-puppeteer": "^3.3.3",
+    "@types/jest": "^24.0.23",
+    "@types/jest-environment-puppeteer": "^4.3.1",
+    "@types/jest-image-snapshot": "^2.11.0",
+    "@types/node": "^12.12.17",
+    "@types/puppeteer": "^2.0.0",
+    "@types/react": "^16.9.16",
+    "@types/react-dom": "^16.9.4",
     "awesome-typescript-loader": "^5.2.1",
-    "fork-ts-checker-webpack-plugin": "^0.5.2",
+    "babel-loader": "^8.0.6",
     "imgur": "^0.3.1",
-    "jest": "^23.6.0",
-    "jest-image-snapshot": "^2.7.0",
-    "jest-puppeteer": "^3.9.0",
-    "puppeteer": "^1.11.0",
-    "start-server-and-test": "^1.7.11",
-    "ts-jest": "^23.10.5",
-    "typescript": "^3.2.4"
+    "jest": "^24.9.0",
+    "jest-image-snapshot": "^2.11.1",
+    "jest-puppeteer": "^4.3.0",
+    "puppeteer": "^2.0.0",
+    "react": "^16.12.0",
+    "react-dom": "^16.12.0",
+    "start-server-and-test": "^1.10.6",
+    "ts-jest": "^24.2.0",
+    "typescript": "^3.7.3"
   },
   "peerDependencies": {
     "react": "^16.3.0-0 || ^17.0.0-0",
@@ -81,7 +81,9 @@
       "^.+\\.tsx?$": "ts-jest"
     },
     "testRegex": "(/__tests__/.*|(\\.|/)(test|spec))\\.[jt]sx?$",
-    "setupTestFrameworkScriptFile": "./setupTests.js",
+    "setupFilesAfterEnv": [
+      "./setupTests.js"
+    ],
     "reporters": [
       "default",
       "<rootDir>/image-reporter.js"
