@@ -1,3 +1,4 @@
+import * as puppeteer from 'puppeteer';
 import { Examples, getTestUrl, getListItems, addFontStyles } from './utils';
 
 jest.setTimeout(10000);
@@ -5,7 +6,7 @@ jest.setTimeout(10000);
 beforeEach(async () => {
   await page.goto(getTestUrl(Examples.BASIC));
   await page.setViewport({ width: 400, height: 800 });
-  await addFontStyles(page);
+  await addFontStyles(page as any);
 });
 
 test('move the first item to second position', async () => {
@@ -13,7 +14,7 @@ test('move the first item to second position', async () => {
   await page.keyboard.press('Space');
   await page.keyboard.press('ArrowDown');
   await page.keyboard.press('Space');
-  expect(await getListItems(page)).toEqual([
+  expect(await getListItems(page as any)).toEqual([
     'Item 2',
     'Item 1',
     'Item 3',
@@ -32,7 +33,7 @@ test('move the sixth item to fifth position', async () => {
   await page.keyboard.press('Space');
   await page.keyboard.press('ArrowUp');
   await page.keyboard.press('Space');
-  expect(await getListItems(page)).toEqual([
+  expect(await getListItems(page as any)).toEqual([
     'Item 1',
     'Item 2',
     'Item 3',
@@ -70,7 +71,7 @@ test('move 1->5, 6->2 and 3->5', async () => {
   await page.keyboard.press('ArrowDown');
   await page.keyboard.press('Space');
 
-  expect(await getListItems(page)).toEqual([
+  expect(await getListItems(page as any)).toEqual([
     'Item 2',
     'Item 6',
     'Item 4',
@@ -87,7 +88,7 @@ test('cancel the move of first item to second position', async () => {
   await page.keyboard.press('Space');
   await page.keyboard.press('ArrowDown');
   await page.keyboard.press('Escape');
-  expect(await getListItems(page)).toEqual([
+  expect(await getListItems(page as any)).toEqual([
     'Item 1',
     'Item 2',
     'Item 3',
