@@ -22,13 +22,13 @@ const POSITIONS = [
 beforeEach(async () => {
   await page.goto(getTestUrl(Examples.BASIC));
   await page.setViewport({ width: 400, height: 800 });
-  await addFontStyles(page);
+  await addFontStyles(page as any);
 });
 
 test('dnd the first item to second position', async () => {
-  await trackMouse(page);
+  await trackMouse(page as any);
   await makeDnd(page.mouse, 1, 2, POSITIONS);
-  expect(await getListItems(page)).toEqual([
+  expect(await getListItems(page as any)).toEqual([
     'Item 2',
     'Item 1',
     'Item 3',
@@ -36,14 +36,14 @@ test('dnd the first item to second position', async () => {
     'Item 5',
     'Item 6'
   ]);
-  await untrackMouse(page);
+  await untrackMouse(page as any);
   expect(await page.screenshot()).toMatchImageSnapshot();
 });
 
 test('dnd the sixth item to fifth position', async () => {
-  await trackMouse(page);
+  await trackMouse(page as any);
   await makeDnd(page.mouse, 6, 5, POSITIONS);
-  expect(await getListItems(page)).toEqual([
+  expect(await getListItems(page as any)).toEqual([
     'Item 1',
     'Item 2',
     'Item 3',
@@ -51,16 +51,16 @@ test('dnd the sixth item to fifth position', async () => {
     'Item 6',
     'Item 5'
   ]);
-  await untrackMouse(page);
+  await untrackMouse(page as any);
   expect(await page.screenshot()).toMatchImageSnapshot();
 });
 
 test('dnd 1->5, 6->2 and 3->5', async () => {
-  await trackMouse(page);
+  await trackMouse(page as any);
   await makeDnd(page.mouse, 1, 5, POSITIONS);
   await makeDnd(page.mouse, 6, 2, POSITIONS);
   await makeDnd(page.mouse, 3, 5, POSITIONS);
-  expect(await getListItems(page)).toEqual([
+  expect(await getListItems(page as any)).toEqual([
     'Item 2',
     'Item 6',
     'Item 4',
@@ -68,6 +68,6 @@ test('dnd 1->5, 6->2 and 3->5', async () => {
     'Item 3',
     'Item 1'
   ]);
-  await untrackMouse(page);
+  await untrackMouse(page as any);
   expect(await page.screenshot()).toMatchImageSnapshot();
 });
