@@ -5,7 +5,7 @@ jest.setTimeout(10000);
 beforeEach(async () => {
   await page.goto(getTestUrl(Examples.HEIGHTS));
   await page.setViewport({ width: 400, height: 800 });
-  await addFontStyles(page);
+  await addFontStyles(page as any);
 });
 
 test('move the first item to second position', async () => {
@@ -13,7 +13,7 @@ test('move the first item to second position', async () => {
   await page.keyboard.press('Space');
   await page.keyboard.press('ArrowDown');
   await page.keyboard.press('Space');
-  expect(await getListItems(page)).toEqual([
+  expect(await getListItems(page as any)).toEqual([
     '100px Item 2',
     '70px Item 1',
     '70px Item 3',
@@ -31,7 +31,7 @@ test('move the fifth item to fourth position', async () => {
   await page.keyboard.press('Space');
   await page.keyboard.press('ArrowUp');
   await page.keyboard.press('Space');
-  expect(await getListItems(page)).toEqual([
+  expect(await getListItems(page as any)).toEqual([
     '70px Item 1',
     '100px Item 2',
     '70px Item 3',
@@ -68,7 +68,7 @@ test('move 1->5, 4->2 and 3->5', async () => {
   await page.keyboard.press('ArrowDown');
   await page.keyboard.press('Space');
 
-  expect(await getListItems(page)).toEqual([
+  expect(await getListItems(page as any)).toEqual([
     '100px Item 2',
     '150px Item 5',
     '70px Item 4',
@@ -84,7 +84,7 @@ test('cancel the move of first item to second position', async () => {
   await page.keyboard.press('Space');
   await page.keyboard.press('ArrowDown');
   await page.keyboard.press('Escape');
-  expect(await getListItems(page)).toEqual([
+  expect(await getListItems(page as any)).toEqual([
     '70px Item 1',
     '100px Item 2',
     '70px Item 3',
