@@ -8,7 +8,7 @@ beforeEach(async () => {
 });
 
 test('scroll down', async () => {
-  await trackMouse(page);
+  await trackMouse(page as any);
   await page.mouse.move(190, 140);
   await page.mouse.down();
   await page.mouse.move(190, 690);
@@ -20,7 +20,7 @@ test('scroll down', async () => {
 });
 
 test('scroll up', async () => {
-  await trackMouse(page);
+  await trackMouse(page as any);
   const list = await page.$('#root ul');
   await page.evaluate(el => (el.scrollTop = 300), list);
   await page.mouse.move(190, 641);
