@@ -33,86 +33,88 @@ const buttonStyles = {
   background: 'transparent'
 };
 
-class Handle extends React.Component<{}, { items: string[] }> {
-  state = {
-    items: ['Item 1', 'Item 2', 'Item 3', 'Item 4', 'Item 5', 'Item 6']
-  };
-  render() {
-    return (
-      <div
-        style={{
-          maxWidth: '300px',
-          margin: '0px auto',
-          backgroundColor: '#F7F7F7',
-          padding: '3em',
-          textAlign: 'center'
-        }}
-      >
-        <List
-          values={this.state.items}
-          onChange={({ oldIndex, newIndex }) =>
-            this.setState((prevState: { items: string[] }) => ({
-              items: arrayMove(prevState.items, oldIndex, newIndex)
-            }))
-          }
-          renderList={({ children, props, isDragged }) => (
-            <ul
-              {...props}
+const Handle: React.FC = () => {
+  const [items, setItems] = React.useState([
+    'Item 1',
+    'Item 2',
+    'Item 3',
+    'Item 4',
+    'Item 5',
+    'Item 6'
+  ]);
+
+  return (
+    <div
+      style={{
+        maxWidth: '300px',
+        margin: '0px auto',
+        backgroundColor: '#F7F7F7',
+        padding: '3em',
+        textAlign: 'center'
+      }}
+    >
+      <List
+        values={items}
+        onChange={({ oldIndex, newIndex }) =>
+          setItems(arrayMove(items, oldIndex, newIndex))
+        }
+        renderList={({ children, props, isDragged }) => (
+          <ul
+            {...props}
+            style={{
+              padding: '0em 0em 1em 0em',
+              cursor: isDragged ? 'grabbing' : 'inherit'
+            }}
+          >
+            {children}
+          </ul>
+        )}
+        renderItem={({ value, props, isDragged, isSelected }) => (
+          <li
+            {...props}
+            style={{
+              ...props.style,
+              padding: '1.5em',
+              margin: '0.5em 0em',
+              listStyleType: 'none',
+              border: '2px solid #CCC',
+              boxShadow: '3px 3px #AAA',
+              color: '#333',
+              borderRadius: '5px',
+              cursor: isDragged ? 'grabbing' : 'inherit',
+              fontFamily: 'Arial, "Helvetica Neue", Helvetica, sans-serif',
+              backgroundColor: isDragged || isSelected ? '#EEE' : '#FFF'
+            }}
+          >
+            <div
               style={{
-                padding: '0em 0em 1em 0em',
-                cursor: isDragged ? 'grabbing' : 'inherit'
+                display: 'flex',
+                alignItems: 'center'
               }}
             >
-              {children}
-            </ul>
-          )}
-          renderItem={({ value, props, isDragged, isSelected }) => (
-            <li
-              {...props}
-              style={{
-                ...props.style,
-                padding: '1.5em',
-                margin: '0.5em 0em',
-                listStyleType: 'none',
-                border: '2px solid #CCC',
-                boxShadow: '3px 3px #AAA',
-                color: '#333',
-                borderRadius: '5px',
-                cursor: isDragged ? 'grabbing' : 'inherit',
-                fontFamily: 'Arial, "Helvetica Neue", Helvetica, sans-serif',
-                backgroundColor: isDragged || isSelected ? '#EEE' : '#FFF'
-              }}
-            >
-              <div
-                style={{
-                  display: 'flex',
-                  alignItems: 'center'
-                }}
-              >
-                {/* 
+              {/* 
                   Mark any node with the data-movable-handle attribute if you wish
                   to use is it as a DnD handle. The rest of renderItem will be then
                   ignored and not start the drag and drop. 
                 */}
-                <button
-                  data-movable-handle
-                  style={{
-                    ...buttonStyles,
-                    cursor: isDragged ? 'grabbing' : 'grab',
-                    marginRight: '3em'
-                  }}
-                  tabIndex={-1}
-                >
-                  <HandleIcon />
-                </button>
-                <div>{value}</div>
-              </div>
-            </li>
-          )}
-        />
-      </div>
-    );
-  }
-}
+              <button
+                data-movable-handle
+                style={{
+                  ...buttonStyles,
+                  cursor: isDragged ? 'grabbing' : 'grab',
+                  marginRight: '3em'
+                }}
+                tabIndex={-1}
+              >
+                <HandleIcon />
+              </button>
+              <div>{value}</div>
+            </div>
+          </li>
+        )}
+      />
+    </div>
+  );
+};
 
 export default Handle;
