@@ -31,100 +31,92 @@ const buttonStyles = {
   background: 'transparent'
 };
 
-class Removable extends React.Component<{}, { items: string[] }> {
-  state = {
-    items: ['Item 1', 'Item 2', 'Item 3', 'Item 4', 'Item 5', 'Item 6']
-  };
-  render() {
-    return (
-      <div
-        style={{
-          maxWidth: '300px',
-          margin: '0px auto',
-          backgroundColor: '#F7F7F7',
-          padding: '3em',
-          textAlign: 'center'
-        }}
-      >
-        <List
-          values={this.state.items}
-          onChange={({ oldIndex, newIndex }) =>
-            this.setState((prevState: { items: string[] }) => ({
-              items: arrayMove(prevState.items, oldIndex, newIndex)
-            }))
-          }
-          renderList={({ children, props, isDragged }) => (
-            <ul
-              {...props}
+const Removable: React.FC = () => {
+  const [items, setItems] = React.useState([
+    'Item 1',
+    'Item 2',
+    'Item 3',
+    'Item 4',
+    'Item 5',
+    'Item 6'
+  ]);
+
+  return (
+    <div
+      style={{
+        maxWidth: '300px',
+        margin: '0px auto',
+        backgroundColor: '#F7F7F7',
+        padding: '3em',
+        textAlign: 'center'
+      }}
+    >
+      <List
+        values={items}
+        onChange={({ oldIndex, newIndex }) =>
+          setItems(arrayMove(items, oldIndex, newIndex))
+        }
+        renderList={({ children, props, isDragged }) => (
+          <ul
+            {...props}
+            style={{
+              padding: '0em 0em 1em 0em',
+              cursor: isDragged ? 'grabbing' : 'inherit'
+            }}
+          >
+            {children}
+          </ul>
+        )}
+        renderItem={({ value, props, index, isDragged, isSelected }) => (
+          <li
+            {...props}
+            style={{
+              ...props.style,
+              padding: '1.5em',
+              margin: '0.5em 0em',
+              listStyleType: 'none',
+              border: '2px solid #CCC',
+              boxShadow: '3px 3px #AAA',
+              color: '#333',
+              borderRadius: '5px',
+              cursor: isDragged ? 'grabbing' : 'grab',
+              fontFamily: 'Arial, "Helvetica Neue", Helvetica, sans-serif',
+              backgroundColor: isDragged || isSelected ? '#EEE' : '#FFF'
+            }}
+          >
+            <div
               style={{
-                padding: '0em 0em 1em 0em',
-                cursor: isDragged ? 'grabbing' : 'inherit'
+                display: 'flex',
+                alignItems: 'center',
+                justifyContent: 'space-between'
               }}
             >
-              {children}
-            </ul>
-          )}
-          renderItem={({ value, props, index, isDragged, isSelected }) => (
-            <li
-              {...props}
-              style={{
-                ...props.style,
-                padding: '1.5em',
-                margin: '0.5em 0em',
-                listStyleType: 'none',
-                border: '2px solid #CCC',
-                boxShadow: '3px 3px #AAA',
-                color: '#333',
-                borderRadius: '5px',
-                cursor: isDragged ? 'grabbing' : 'grab',
-                fontFamily: 'Arial, "Helvetica Neue", Helvetica, sans-serif',
-                backgroundColor: isDragged || isSelected ? '#EEE' : '#FFF'
-              }}
-            >
-              <div
-                style={{
-                  display: 'flex',
-                  alignItems: 'center',
-                  justifyContent: 'space-between'
+              <div>{value}</div>{' '}
+              <button
+                onClick={() => {
+                  setItems(
+                    typeof index !== 'undefined'
+                      ? arrayRemove(items, index)
+                      : items
+                  );
                 }}
+                style={buttonStyles}
               >
-                <div>{value}</div>{' '}
-                <button
-                  onClick={() => {
-                    this.setState(prevProps => ({
-                      items:
-                        typeof index !== 'undefined'
-                          ? arrayRemove(prevProps.items, index)
-                          : prevProps.items
-                    }));
-                  }}
-                  style={buttonStyles}
-                >
-                  <RemovableIcon />
-                </button>
-              </div>
-            </li>
-          )}
-        />
-        <button
-          onClick={() =>
-            this.setState({
-              items: [
-                'Item 1',
-                'Item 2',
-                'Item 3',
-                'Item 4',
-                'Item 5',
-                'Item 6'
-              ]
-            })
-          }
-        >
-          Reset
-        </button>
-      </div>
-    );
-  }
-}
+                <RemovableIcon />
+              </button>
+            </div>
+          </li>
+        )}
+      />
+      <button
+        onClick={() =>
+          setItems(['Item 1', 'Item 2', 'Item 3', 'Item 4', 'Item 5', 'Item 6'])
+        }
+      >
+        Reset
+      </button>
+    </div>
+  );
+};
 
 export default Removable;
