@@ -1,15 +1,6 @@
 import * as React from 'react';
 import { List, arrayMove } from '../src/index';
 
-interface IAppStateItems {
-  items: {
-    first_name: string;
-    last_name: string;
-    car_make: string;
-    car_model: string;
-  }[];
-}
-
 const tableStyles = {
   background: '#eaebec',
   borderSpacing: 0
@@ -33,115 +24,110 @@ const tdStyles = {
   width: '150px'
 } as React.CSSProperties;
 
-class Table extends React.Component<{}, IAppStateItems> {
-  state = {
-    items: [
-      {
-        first_name: 'Oralie',
-        last_name: 'Blaszkiewicz',
-        car_make: 'Volkswagen',
-        car_model: 'Eurovan'
-      },
-      {
-        first_name: 'Marylin',
-        last_name: 'Seagar',
-        car_make: 'BMW',
-        car_model: 'X3'
-      },
-      {
-        first_name: 'Cristy',
-        last_name: 'Carberry',
-        car_make: 'Chevrolet',
-        car_model: 'Camaro'
-      },
-      {
-        first_name: 'Oliviero',
-        last_name: 'Methven',
-        car_make: 'Chevrolet',
-        car_model: 'Impala'
-      },
-      {
-        first_name: 'Eduardo',
-        last_name: 'Rowan',
-        car_make: 'Mercedes-Benz',
-        car_model: 'M-Class'
-      },
-      {
-        first_name: 'Georgianne',
-        last_name: 'Rainville',
-        car_make: 'Mitsubishi',
-        car_model: 'Mirage'
-      },
-      {
-        first_name: 'Cristi',
-        last_name: 'Kollach',
-        car_make: 'Cadillac',
-        car_model: 'Seville'
-      }
-    ]
-  };
-  render() {
-    return (
-      <div
-        style={{
-          padding: '3em',
-          display: 'flex',
-          justifyContent: 'center'
-        }}
-      >
-        <List
-          values={this.state.items}
-          onChange={({ oldIndex, newIndex }) =>
-            this.setState((prevState: IAppStateItems) => ({
-              items: arrayMove(prevState.items, oldIndex, newIndex)
-            }))
-          }
-          renderList={({ children, props, isDragged }) => (
-            <table
+const Table: React.FC = () => {
+  const [items, setItems] = React.useState([
+    {
+      first_name: 'Oralie',
+      last_name: 'Blaszkiewicz',
+      car_make: 'Volkswagen',
+      car_model: 'Eurovan'
+    },
+    {
+      first_name: 'Marylin',
+      last_name: 'Seagar',
+      car_make: 'BMW',
+      car_model: 'X3'
+    },
+    {
+      first_name: 'Cristy',
+      last_name: 'Carberry',
+      car_make: 'Chevrolet',
+      car_model: 'Camaro'
+    },
+    {
+      first_name: 'Oliviero',
+      last_name: 'Methven',
+      car_make: 'Chevrolet',
+      car_model: 'Impala'
+    },
+    {
+      first_name: 'Eduardo',
+      last_name: 'Rowan',
+      car_make: 'Mercedes-Benz',
+      car_model: 'M-Class'
+    },
+    {
+      first_name: 'Georgianne',
+      last_name: 'Rainville',
+      car_make: 'Mitsubishi',
+      car_model: 'Mirage'
+    },
+    {
+      first_name: 'Cristi',
+      last_name: 'Kollach',
+      car_make: 'Cadillac',
+      car_model: 'Seville'
+    }
+  ]);
+
+  return (
+    <div
+      style={{
+        padding: '3em',
+        display: 'flex',
+        justifyContent: 'center'
+      }}
+    >
+      <List
+        values={items}
+        onChange={({ oldIndex, newIndex }) =>
+          setItems(arrayMove(items, oldIndex, newIndex))
+        }
+        renderList={({ children, props, isDragged }) => (
+          <table
+            style={{
+              ...tableStyles,
+              cursor: isDragged ? 'grabbing' : undefined
+            }}
+          >
+            <thead>
+              <tr>
+                <th style={thStyles}>First name</th>
+                <th style={thStyles}>Last name</th>
+                <th style={thStyles}>Car maker</th>
+                <th style={thStyles}>Car model</th>
+              </tr>
+            </thead>
+            <tbody {...props}>{children}</tbody>
+          </table>
+        )}
+        renderItem={({ value, props, isDragged, isSelected }) => {
+          const row = (
+            <tr
+              {...props}
               style={{
-                ...tableStyles,
-                cursor: isDragged ? 'grabbing' : undefined
+                ...props.style,
+                cursor: isDragged ? 'grabbing' : 'grab',
+                backgroundColor: isDragged || isSelected ? '#EEE' : '#fafafa'
               }}
             >
-              <thead>
-                <tr>
-                  <th style={thStyles}>First name</th>
-                  <th style={thStyles}>Last name</th>
-                  <th style={thStyles}>Car maker</th>
-                  <th style={thStyles}>Car model</th>
-                </tr>
-              </thead>
-              <tbody {...props}>{children}</tbody>
+              <td style={tdStyles}>{value.first_name}</td>
+              <td style={tdStyles}>{value.last_name}</td>
+              <td style={tdStyles}>{value.car_make}</td>
+              <td style={tdStyles}>{value.car_model}</td>
+            </tr>
+          );
+          return isDragged ? (
+            <table style={{ ...props.style, borderSpacing: 0 }}>
+              <tbody>{row}</tbody>
             </table>
-          )}
-          renderItem={({ value, props, isDragged, isSelected }) => {
-            const row = (
-              <tr
-                {...props}
-                style={{
-                  ...props.style,
-                  cursor: isDragged ? 'grabbing' : 'grab',
-                  backgroundColor: isDragged || isSelected ? '#EEE' : '#fafafa'
-                }}
-              >
-                <td style={tdStyles}>{value.first_name}</td>
-                <td style={tdStyles}>{value.last_name}</td>
-                <td style={tdStyles}>{value.car_make}</td>
-                <td style={tdStyles}>{value.car_model}</td>
-              </tr>
-            );
-            return isDragged ? (
-              <table style={{ ...props.style, borderSpacing: 0 }}>
-                <tbody>{row}</tbody>
-              </table>
-            ) : (
-              row
-            );
-          }}
-        />
-      </div>
-    );
-  }
-}
+          ) : (
+            row
+          );
+        }}
+      />
+    </div>
+  );
+};
 
 export default Table;
